requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cachetools==5.5.2
certifi==2025.7.9
charset-normalizer==3.4.2
cloudinary==1.44.1
Django==3.2.25
django-cloudinary-storage==0.3.0
django-cors-headers==4.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.10
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
httplib2==0.22.0
idna==3.10
inflection==0.5.1
oauthlib==3.2.2
packaging==24.0
Pillow==9.5.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.4
python-dotenv==0.21.1
pytz==2025.2
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.4.4
typing-extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.18
